requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
PyJWT
websockets

